requirements.txt
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-ollama>=0.1.0
langchain-mistralai>=0.1.0
langgraph>=0.1.0
chromadb>=0.4.0
google-generativeai>=0.3.0
google-cloud-storage>=2.10.0
slack-bolt>=1.18.0
Pillow>=10.0.0 
jira>=3.0.0